# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexora AI Business OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: getTickets
      tags: [tickets]
      summary: List tickets
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: priority
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
...
  /hr/employees:
    get:
      operationId: getEmployees
      tags: [hr]
      summary: List employees
      parameters:
        - name: department
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [hr]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /hr/employees/{id}:
    patch:
      operationId: updateEmployee
      tags: [hr]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
...
  /omnichannel/conversations:
    get:
      operationId: getConversations
      tags: [omnichannel]
      summary: List conversations
      parameters:
        - name: channel
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
  /omnichannel/conversations/{id}/messages:
    get:
      operationId: getMessages
      tags: [omnichannel]
      summary: Get messages in conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [omnichannel]
      summary: Send message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # AI Brain
...
components:
  schemas:
...
    Ticket:
      type: object
      required: [id, title, status, priority, category, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        category:
          type: string
        assignedTo:
          type: ["string", "null"]
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        slaDeadline:
          type: ["string", "null"]
        slaStatus:
          type: ["string", "null"]
        aiSuggestedResponse:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    TicketInput:
      type: object
      required: [title, priority, category]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
        category:
          type: string
        assignedTo:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
    TicketUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        assignedTo:
          type: ["string", "null"]
        slaStatus:
          type: ["string", "null"]
...
    Employee:
      type: object
      required: [id, name, email, department, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        department:
          type: string
        role:
          type: string
        status:
          type: string
        productivity:
          type: ["number", "null"]
        goalsCompleted:
          type: ["integer", "null"]
        goalsTotal:
          type: ["integer", "null"]
        points:
          type: ["integer", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    EmployeeInput:
      type: object
      required: [name, email, department, role]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        department:
          type: string
        role:
          type: string
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        department:
          type: string
        role:
          type: string
        status:
          type: string
        productivity:
          type: ["number", "null"]
...
    Conversation:
      type: object
      required: [id, channel, customerName, status, lastMessage, lastMessageAt]
      properties:
        id:
          type: integer
        channel:
          type: string
        customerName:
          type: string
        customerAvatar:
          type: ["string", "null"]
        status:
          type: string
        lastMessage:
          type: string
        lastMessageAt:
          type: string
        unreadCount:
          type: integer
        assignedTo:
          type: ["string", "null"]
    Message:
      type: object
      required: [id, conversationId, content, direction, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        content:
          type: string
        direction:
          type: string
        senderName:
          type: ["string", "null"]
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # AI
...